delete,add
Deleted faker library since we will not need that library. Custom data will be created by custom functions in test classes to provide migration sql files to insert that data in postgre database. Created  data entry class with java doc and function to create a list of sex which depends on number of records. This function is not completed, since we dont have a random function to generate male or female.