added
added function which make script file  which inserts data from sex and countries into the database. Not completed. We must make a two queries and fetch data from two lists, it  is better to do one list per time. Then we define a query in the list, fetching the values and then saving it to the file which is script which add  some data into table. Other data such as adress contact and city we need to add manually, exception is name and surname, we can create some fake data using faker. Or we can write our own functions.