skeleton done
skeleton of this application is done. we only need to limit user rights for example user maybe could insert sex but not to update them or delete them users would have a right only to use crud complete operations for contacts. Users could only insert data to other tables but not  to edit them or delete them. Admin role would have that access. Second thing, we need to implement that security. We only have one jsp page, we will add another one for login. We need to think a way how to limit access to one page crud operations on some roles.