added
added new faker version also added test class in which you enter a value how many vusers you want to generate and it simply function uses faker to poulate data except email  and roles role admin is given every person if number in a loop is not add and also the aga if number is add give the age value if not value is null it is also generated an id, which is not needed  but it is made just becaue of cosmetics. To look nicer in console.